[Backup] Ktorからjava.net.http.HttpClientに移行。超軽量化

+++ SS-Backup/src/main/kotlin/com/github/syari/ss/plugins/backup/WebDAVUploader.kt
@@ -1,19 +1,15 @@
 package com.github.syari.ss.plugins.backup
 
 import com.github.syari.ss.plugins.backup.Main.Companion.plugin
-import io.ktor.client.HttpClient
-import io.ktor.client.content.LocalFileContent
-import io.ktor.client.engine.okhttp.OkHttp
-import io.ktor.client.features.auth.Auth
-import io.ktor.client.features.auth.providers.basic
-import io.ktor.client.request.request
-import io.ktor.client.statement.HttpResponse
-import io.ktor.http.HttpMethod
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
+import com.github.syari.ss.plugins.backup.http.buildHttpClient
+import com.github.syari.ss.plugins.backup.http.buildHttpRequest
+import com.github.syari.ss.plugins.core.scheduler.CreateScheduler.run
 import java.io.File
-import java.net.SocketException
+import java.net.http.HttpClient
+import java.net.http.HttpRequest
+import java.net.http.HttpResponse
+import java.time.Duration
+import java.util.Base64
 
 class WebDAVUploader(private val url: String, user: String, pass: String) {
     companion object {
@@ -28,37 +24,26 @@ class WebDAVUploader(private val url: String, user: String, pass: String) {
         }
     }
 
-    private val client = HttpClient(OkHttp) {
-        install(Auth) {
-            basic {
-                username = user
-                password = pass
-            }
-        }
-        engine {
-            config {
-                retryOnConnectionFailure(true)
-            }
-        }
+    private val client = buildHttpClient {
+        followRedirects(HttpClient.Redirect.NORMAL)
+        connectTimeout(Duration.ofSeconds(10))
     }
 
-    fun upload(file: File) {
-        GlobalScope.launch {
-            repeat(10) {
-                delay(100)
-                try {
-                    client.request<HttpResponse>("$url/${file.name}") {
-                        method = HttpMethod.Put
-                        body = LocalFileContent(file)
-                    }
-                    plugin.logger.info("${file.path} をアップロードしました")
-                    file.delete()
-                    return@launch
-                } catch (ex: SocketException) {
+    private val authBase64 = Base64.getEncoder().encodeToString("$user:$pass".toByteArray())
 
-                }
+    fun upload(file: File) {
+        run(plugin, async = true) {
+            val response: HttpResponse<String> = client.send(buildHttpRequest("$url/${file.name}") {
+                header("Authorization", "Basic $authBase64")
+                PUT(HttpRequest.BodyPublishers.ofFile(file.toPath()))
+            }, HttpResponse.BodyHandlers.ofString())
+            val statusCode = response.statusCode()
+            if (statusCode in 200 until 300) {
+                plugin.logger.info("${file.path} をアップロードしました")
+                file.delete()
+            } else {
+                plugin.logger.warning("${file.path} のアップロードに失敗しました ($statusCode)")
             }
-            plugin.logger.severe("${file.path} のアップロードに失敗しました")
         }
     }
 }
